Reset service-mirror component when target's k8s API is unreachable
(This is based off of the `alpeb/remove-multicluster-install-test` branch, followup to #4995 and is another preliminary required for the upcoming multicluster integration test).

When the service-mirror component can't reach the target's k8s API, the goroutine blocks and it can't be unblocked.

This was happenining specifically  in the case of the multicluster integration test (still to be pushed), where the source and target clusters are created in quick succession and the target's API service doesn't always have time to be exposed before being requested by the service mirror.

The fix consists on no longer have `restartClusterWatcher` be side-effecting, and instead return an error. If such error is not nil then the link watcher is stopped and reset after 10 seconds.