Multicluster integration test
(This is a followup to #4997 and it's based off of its branch `alpeb/certs-openssl`)

This implements the `run_multicluster_test()` function in `bin/_test-helpers.sh`.

The idea is to create two clusters (source and target) using k3d, with linkerd and multicluster support in both, plus emojivoto (without vote-bot) in target and vote-bot in source.
We then link the clusters and make sure traffic is flowing.

Detailed sequence:
- Create certficates.
- Install linkerd along with multicluster support in the target cluster.
- Run the `target1` test: install emojivoto in the target cluster (without vote-bot).
- Run `linkerd mc link` on the target cluster.
- Install linkerd along with multicluster support in the source cluster.
- Apply the `link` resource in the source cluster.
- Run the `source` test: Check `linkerd mc gateways` returns the target cluster link, and only install emojivoto's vote-bot in the source cluster.
- Run the `target2` test: Make sure `web-svc` in the target cluster is receiving requests.